Refactor Category components for improved styling and structure
- Update import paths for CSS modules in CategoryContent and ChildCategoryCard
- Enhance ChildCategoryCard layout with new styles and animations
- Modify CategoryContent to display a default description when none is provided
