+++ src/components/V2/category/CategoryContent.module.css
@@ -1,0 +1,334 @@
+/* Premium modern container */
+.container {
+  max-width: 1400px;
+  margin: 0 auto;
+  padding: 80px 20px;
+  background: linear-gradient(to bottom, #f8fafc 0%, #ffffff 100%);
+  position: relative;
+  overflow: hidden;
+}
+
+.container::before {
+  content: '';
+  position: absolute;
+  top: 0;
+  left: 0;
+  width: 100%;
+  height: 100%;
+  background-image: radial-gradient(
+      circle at 20% 10%,
+      rgba(64, 81, 181, 0.03) 0%,
+      transparent 30%
+    ),
+    radial-gradient(
+      circle at 80% 80%,
+      rgba(26, 35, 126, 0.03) 0%,
+      transparent 30%
+    );
+  pointer-events: none;
+  z-index: 0;
+}
+
+/* Header styling */
+.header {
+  text-align: center;
+  position: relative;
+  margin-bottom: 60px;
+  padding: 0 20px;
+  z-index: 1;
+}
+
+.headerAccent {
+  width: 120px;
+  height: 4px;
+  background: linear-gradient(90deg, #1a237e 0%, #4051b5 100%);
+  margin: 0 auto 20px;
+  border-radius: 50px;
+}
+
+.title {
+  font-size: 44px;
+  font-weight: 800;
+  color: #1a237e;
+  margin-bottom: 16px;
+  position: relative;
+  display: inline-block;
+  letter-spacing: -0.02em;
+  line-height: 1.2;
+}
+
+.title::after {
+  content: '';
+  position: absolute;
+  bottom: -6px;
+  left: 50%;
+  transform: translateX(-50%);
+  width: 60%;
+  height: 4px;
+  background: linear-gradient(90deg, #1a237e 30%, #4051b5 100%);
+  border-radius: 2px;
+}
+
+.description {
+  font-size: 18px;
+  color: #64748b;
+  max-width: 600px;
+  margin: 20px auto 0;
+  line-height: 1.6;
+}
+
+.divider {
+  display: none;
+}
+
+/* Grid container */
+.gridContainer {
+  position: relative;
+  z-index: 1;
+  padding: 0;
+}
+
+.grid {
+  display: grid;
+  grid-template-columns: repeat(auto-fill, minmax(300px, 1fr));
+  gap: 30px;
+  margin: 0 auto;
+  max-width: 1300px;
+}
+
+/* Card styling for ChildCategoryCard */
+.categoryCard {
+  background: #ffffff;
+  border-radius: 16px;
+  box-shadow: 0 10px 25px -5px rgba(0, 0, 0, 0.05);
+  overflow: hidden;
+  transition: all 0.4s cubic-bezier(0.175, 0.885, 0.32, 1.275);
+  height: 100%;
+  display: flex;
+  flex-direction: column;
+  padding: 0;
+  position: relative;
+  width: 100%;
+  border: 1px solid rgba(203, 213, 225, 0.5);
+  transform: translateY(0);
+}
+
+.categoryCard:hover {
+  transform: translateY(-10px);
+  box-shadow: 0 20px 30px -10px rgba(0, 0, 0, 0.1);
+  border-color: rgba(26, 35, 126, 0.2);
+}
+
+.cardInner {
+  display: flex;
+  flex-direction: column;
+  height: 100%;
+  padding: 30px;
+  position: relative;
+  z-index: 2;
+}
+
+.cardTop {
+  height: 8px;
+  background: linear-gradient(90deg, #1a237e, #4051b5);
+  position: absolute;
+  top: 0;
+  left: 0;
+  right: 0;
+  z-index: 1;
+}
+
+.cardDecoration {
+  position: absolute;
+  width: 140px;
+  height: 140px;
+  background: linear-gradient(
+    135deg,
+    rgba(26, 35, 126, 0.04),
+    rgba(64, 81, 181, 0.06)
+  );
+  top: -70px;
+  right: -70px;
+  border-radius: 50%;
+  transition: all 0.5s ease;
+  opacity: 0.5;
+  z-index: 1;
+}
+
+.categoryCard:hover .cardDecoration {
+  transform: scale(1.2);
+  opacity: 0.8;
+}
+
+.categoryCard:nth-child(even) .cardDecoration {
+  top: auto;
+  right: auto;
+  bottom: -70px;
+  left: -70px;
+}
+
+.categoryTitle {
+  font-size: 24px;
+  font-weight: 700;
+  color: #0f172a;
+  margin-bottom: 16px;
+  transition: color 0.3s ease;
+  line-height: 1.3;
+  position: relative;
+  z-index: 2;
+}
+
+.categoryCard:hover .categoryTitle {
+  color: #1a237e;
+}
+
+/* Category Meta section - updated for single element */
+.categoryMeta {
+  display: flex;
+  align-items: center;
+  justify-content: flex-end;
+  margin-top: auto;
+  padding-top: 20px;
+  border-top: 1px solid #f1f5f9;
+  position: relative;
+  z-index: 2;
+}
+
+/* Shop Now button - enhanced for better visibility */
+.shopNow {
+  color: #1a237e;
+  font-weight: 600;
+  font-size: 15px;
+  display: flex;
+  align-items: center;
+  transition: all 0.3s ease;
+  padding: 10px 20px;
+  border-radius: 30px;
+  background: rgba(26, 35, 126, 0.08);
+}
+
+.categoryCard:hover .shopNow {
+  background: rgba(26, 35, 126, 0.15);
+  transform: translateX(4px);
+}
+
+.arrowIcon {
+  margin-left: 8px;
+  transition: transform 0.3s ease;
+}
+
+.categoryCard:hover .arrowIcon {
+  transform: translateX(4px);
+}
+
+/* Animation for cards */
+.fadeIn {
+  animation: fadeInUp 0.5s forwards;
+  opacity: 0;
+}
+
+@keyframes fadeInUp {
+  from {
+    opacity: 0;
+    transform: translateY(20px);
+  }
+  to {
+    opacity: 1;
+    transform: translateY(0);
+  }
+}
+
+/* Count badge */
+.countBadge {
+  display: inline-flex;
+  align-items: center;
+  justify-content: center;
+  background-color: #f1f5f9;
+  color: #475569;
+  font-size: 13px;
+  font-weight: 600;
+  height: 24px;
+  min-width: 24px;
+  padding: 0 8px;
+  border-radius: 12px;
+  margin-left: 8px;
+  transition: all 0.3s ease;
+}
+
+.categoryCard:hover .countBadge {
+  background-color: #1a237e;
+  color: white;
+}
+
+/* Media queries for responsive design */
+@media (max-width: 1200px) {
+  .title {
+    font-size: 40px;
+  }
+
+  .description {
+    font-size: 17px;
+  }
+}
+
+@media (max-width: 767px) {
+  .container {
+    padding: 60px 15px;
+  }
+
+  .header {
+    margin-bottom: 40px;
+  }
+
+  .title {
+    font-size: 32px;
+  }
+
+  .description {
+    font-size: 16px;
+  }
+
+  .grid {
+    grid-template-columns: repeat(auto-fill, minmax(280px, 1fr));
+    gap: 20px;
+  }
+
+  .cardInner {
+    padding: 25px;
+  }
+
+  .categoryTitle {
+    font-size: 22px;
+  }
+}
+
+@media (max-width: 520px) {
+  .container {
+    padding: 40px 15px;
+  }
+
+  .title {
+    font-size: 28px;
+  }
+
+  .description {
+    font-size: 15px;
+  }
+
+  .grid {
+    grid-template-columns: 1fr;
+  }
+
+  .cardInner {
+    padding: 20px;
+  }
+
+  .categoryTitle {
+    font-size: 20px;
+  }
+
+  .shopNow {
+    font-size: 14px;
+    padding: 6px 12px;
+  }
+}
